+++ src/Firebase/Messaging/WebPushConfig.php
@@ -121,13 +121,11 @@ final class WebPushConfig implements JsonSerializable
             }
         }
 
-        if (array_key_exists('Urgency', $headers)) {
-            if (!in_array($headers['Urgency'], self::VALID_URGENCIES, true)) {
-                throw new InvalidArgument(sprintf(
-                    'The Urgency in the WebPushConfig header must must be one of %s',
-                    implode(',', self::VALID_URGENCIES)
-                ));
-            }
+        if (array_key_exists('Urgency', $headers) && !in_array($headers['Urgency'], self::VALID_URGENCIES, true)) {
+            throw new InvalidArgument(sprintf(
+                'The Urgency in the WebPushConfig header must must be one of %s',
+                implode(',', self::VALID_URGENCIES)
+            ));
         }
 
         return $headers;