+++ tests/Integration/Database/AuthVariableOverrideTest.php
@@ -12,6 +12,8 @@ use Kreait\Firebase\Tests\Integration\DatabaseTestCase;
 
 /**
  * @internal
+ * @group database-emulator
+ * @group emulator
  */
 final class AuthVariableOverrideTest extends DatabaseTestCase
 {