+++ tests/Integration/Database/RuleSetTest.php
@@ -10,6 +10,8 @@ use Kreait\Firebase\Tests\Integration\DatabaseTestCase;
 
 /**
  * @internal
+ * @group database-emulator
+ * @group emulator
  */
 final class RuleSetTest extends DatabaseTestCase
 {