+++ tests/Integration/DatabaseTest.php
@@ -9,11 +9,17 @@ use Kreait\Firebase\Factory;
 
 /**
  * @internal
+ * @group database-emulator
+ * @group emulator
  */
 final class DatabaseTest extends DatabaseTestCase
 {
     public function testWithNonExistingDatabase(): void
     {
+        if (self::databaseIsEmulated()) {
+            $this->markTestSkipped('The RTDB emulator creates databases if they don\'t exist');
+        }
+
         $credentials = self::$serviceAccount->asArray();
         $credentials['project_id'] = 'non-existing';